Adds a few data-process helpers and basic graphs
After reading things like the article at the bottom of this commit and
realizing that to _some_ extent processing data in the browser is
necessary, I figured I might as well build a few group helpers that
take a simple array of data and convert them into formats that are
usable by victory.

Based on a back of envelope analysis, it looks like this can be done for
datasets of a few hundred items at ~60 fps or less.

https://towardsdatascience.com/advanced-visualization-with-react-vis-efc5c6667b4